+++ packages/apollo-utilities/CHANGELOG.md
@@ -1,8 +1,8 @@
 # Change log
 
-
 ### vNext
 - Map coverage to original source
+- Added ability to include or ignore fragments in `getDirectivesFromDocument` [PR#3010](https://github.com/apollographql/apollo-client/pull/3010)
 
 ### 1.1.3
 - dependency updates